requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncclick==8.1.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-datetime==0.2
cryptography==42.0.8
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyHS100==0.3.5.2
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-kasa @ git+https://github.com/python-kasa/python-kasa.git@4cf395483f7a8a51e47bc7ece150d2da26de57fe
pytz==2024.1
requests==2.32.3
selenium==4.24.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.32
tplink-cloud-api==4.1.5
trio==0.26.2
trio-websocket==0.11.1
twilio==9.2.4
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.43.0
wsproto==1.2.0
yarl==1.9.4
